switching to react router
