+++ package.json
@@ -9,6 +9,8 @@
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.0"
   },
   "scripts": {
