+++ src/components/Buttons/CalendarDatesButton/CalendarDatesButton.js
@@ -1,10 +1,13 @@
 import React from 'react';
 import classes from './CalendarDatesButton.module.css';
 
-
 const CalendarDatesButton = props => {
   return (
-    <a aria-label={'March '+ props.date} className={classes.calendarButtonStyling} href='#'>
+    <a
+      aria-label={'March ' + props.date}
+      className={classes.calendarButtonStyling}
+      href='/'
+    >
       {props.date}
     </a>
   );
